requirements.txt
uvicorn
fastapi
fastapi-sqlalchemy
psycopg2-binary
alembic
pydantic[dotenv]
pyjwt
